+++ CHANGES.rst
@@ -10,14 +10,17 @@ Breaking changes:
 
 New features:
 
+- Add display template for AjaxSelectWidget showing the actual vocabularies term title.
+  [jensens]
+
+Bug fixes:
+
 - Fixes AjaxSelectWidget to respect tokens different from values in vocabularies.
   This includes changes in both, the converter and the widget itself.
   A test was added too.
   ``get_ajaxselect_options`` from ``plone.app.widgets.utils`` is assimilated by the widget now too simplify the whole code,
   so the one in the other package is dead code now and will be deprecated there.
   [jensens]
-
-Bug fixes:
 
 - LinkFieldWidget: added converter method toFieldValue [ksuess]